front end updates

+++ src/main/java/com/prodemy/controller/AdminController.java
@@ -7,15 +7,13 @@ import org.springframework.scheduling.config.Task;
 import org.springframework.security.access.prepost.*;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.*;
-
 
 @RestController
 @RequestMapping("/api/admin")
 @PreAuthorize("hasRole('ADMIN')")  // Spring Security annotation
 public class AdminController {
     @PostMapping("/ban-user/{userId}")
-    public ResponseEntity<?> banUser(@PathVariable Long userId) {
+    public ResponseEntity<?> banUser(@PathVariable int userId) {
         // Logic to ban a user
     }
 }
